// src/common/enums/grade.enum.ts
export enum Grade {
  PRE_SCHOOL = 'Pre-School',
  KINDERGARTEN = 'Kindergarten',
  GRADE_1 = 'Grade 1',
  GRADE_2 = 'Grade 2',
  GRADE_3 = 'Grade 3',
  GRADE_4 = 'Grade 4',
  GRADE_5 = 'Grade 5',
  GRADE_6 = 'Grade 6',
  GRADE_7 = 'Grade 7',
  GRADE_8 = 'Grade 8',
  GRADE_9 = 'Grade 9',
  GRADE_10 = 'Grade 10',
  GRADE_11 = 'Grade 11',
  GRADE_12 = 'Grade 12',
  O_LEVEL = 'O-Level',
  A_LEVEL = 'A-Level',
  MATRICULATION = 'Matriculation',
  INTERMEDIATE = 'Intermediate',
  BACHELORS = 'Bachelors',
  MASTERS = 'Masters',
  PHD = 'PhD',
}