// src/common/enums/interest.enum.ts
export enum Interest {
  MATHEMATICS = 'Mathematics',
  SCIENCE = 'Science',
  ENGLISH = 'English',
  COMPUTER_SCIENCE = 'Computer Science',
  HISTORY = 'History',
  GEOGRAPHY = 'Geography',
  LITERATURE = 'Literature',
  LANGUAGES = 'Languages',
  PHYSICAL_EDUCATION = 'Physical Education',
  ECONOMICS = 'Economics',
  BUSINESS = 'Business',
  PSYCHOLOGY = 'Psychology',
  PHILOSOPHY = 'Philosophy',
  CHEMISTRY = 'Chemistry',
  BIOLOGY = 'Biology',
  PHYSICS = 'Physics',
  ACCOUNTING = 'Accounting',
  STATISTICS = 'Statistics',
  ART = 'Art',
  MUSIC = 'Music',
  SOCIOLOGY = 'Sociology',
  POLITICAL_SCIENCE = 'Political Science',
  LAW = 'Law',
  MEDICAL = 'Medical Science',
  ENGINEERING = 'Engineering',
  COMPUTER_ENGINEERING = 'Computer Engineering',
  ARCHITECTURE = 'Architecture',
  AGRICULTURE = 'Agriculture',
  PSYCHOLINGUISTICS = 'Psycholinguistics',
  RELIGION = 'Religion Studies',
  ENVIRONMENTAL_SCIENCE = 'Environmental Science',
  ENTREPRENEURSHIP = 'Entrepreneurship',
}